[refector] SocketMessage관련 로직 레이어 구분
